requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docopt==0.6.2
exceptiongroup==1.2.0
Flask==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.1
langchain-google-genai==0.0.5
langsmith==0.0.72
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
pipreqs==0.4.13
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
yarg==0.1.9
